fix: use timezone aware timezones from db, fix dashboard api
